backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
supabase==2.5.1
pydantic==2.7.1
pydantic-settings==2.2.1
pydantic[email]==2.7.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
cloudinary
